Updates SAM presets.
For byrack grouping the groupsize is 0 instead of 1.

Note updates.

Fix HBASE groups changed to 0 instead of 1.

If you want to do grouping totatly by rack you use 0 for the groupsize. If you want to use the logic but still have a certain limit to how many in parallel they you use the groupsize value of 1 or more.